Fixing inconsistencies on replacesStatements
The guidance was inconsistent in it's reference to `replacesStatement` vs `replacesStatements` and the language suggesting that only one previous statement could be referred to.

I've fixed this. As guidance was previously outside of the normative parts of the documentation, I believe this fix should be possible as a 'typo' fix within the layout updates to schema-beta-2.