Update transifex-client to fix RTD builds
BODS 0.2.0 no longer builds on master due to incompatibilties between
the requirements of transifex-client @ 0.13.6 and our pinned version of
urllib3.

Updating transifex-client on it's own seems to resolve the build
problem, so in the interests of changing this as little as possible I've
not updated anything else, even where we're using a newer version of
other deps in master.